// application-gateway-typescript/src/types.ts
// Path to org1 user private key directory.
export type Order = {
    bidMatchId: number;
    bidStatus: number;
    id: number;
    onMarketPrice: number;
    orderCost: number;
    paymentId: number;
    slotId: string;
    slotExecDate: number;
    totalQuantity: number;
    unitCost: number;
    action: number;
    userId: number;
};